+++ src/Core/OnionPronia.Domain/Entities/ProductTag.cs
@@ -8,8 +8,8 @@ namespace OnionPronia.Domain.Entities
 {
     public class ProductTag
     {
-        public int ProductId { get; set; }
-        public int TagId { get; set; }
+        public long ProductId { get; set; }
+        public long TagId { get; set; }
         public Product Product { get; set; }
         public Tag Tag { get; set; }
     }
